+++ tests/conftest.py
@@ -1,6 +1,7 @@
 # Some utility functions for testing
 import json
 from pathlib import Path
+from .evaluation.helper import get_html, read_or_download_json
 
 
 def get_data(filename, resource_type):
@@ -24,3 +25,21 @@ def get_data(filename, resource_type):
     else:
         with open(file.with_suffix(".json"), "r", encoding="utf-8") as f:
             return json.load(f)
+
+
+def get_scrapinghub_data(filename: str):
+    ground_truth_json = read_or_download_json(
+        "https://raw.githubusercontent.com/scrapinghub/article-extraction-benchmark/master/ground-truth.json"
+    )
+    if filename not in ground_truth_json:
+        raise ValueError(f"Invalid filename {filename}")
+
+    html_content = get_html(
+        f"https://github.com/scrapinghub/article-extraction-benchmark/raw/master/html/{filename}.html.gz"
+    )
+
+    return {
+        "url": ground_truth_json[filename]["url"],
+        "html": html_content,
+        "text": ground_truth_json[filename]["articleBody"],
+    }
