+++ tests/data/metadata/article_with_divs.json
@@ -21,21 +21,12 @@
     ],
     "movies": [],
     "keywords": [
-        "photos",
-        "actress",
         "kate",
         "henshaw",
-        "stunning",
+        "actress",
         "nollywood",
-        "media",
-        "nigerian",
-        "released",
-        "online",
-        "making",
-        "rounds",
-        "social",
-        "absolutely",
-        "make-up",
+        "stunning",
+        "photos",
         "veteran",
         "2008",
         "won",
@@ -52,10 +43,16 @@
         "hand",
         "bag",
         "colours",
+        "nigerian",
         "flag",
         "47",
         "guests",
-        "invited"
+        "invited",
+        "event",
+        "media",
+        "mogul",
+        "debola",
+        "lagos"
     ],
     "meta_keywords": [
         "Nigeria breaking news",
@@ -74,12 +71,12 @@
         "LinkNaija"
     ],
     "publish_date": "2018-10-06T08:57:00+02:00",
-    "summary": "Photos which were released online and are making the rounds on social media, show Nigerian actress, Kate Henshaw, looking absolutely stunning in new make-up photos.\nKate Henshaw Veteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.\nShe flaunted her hand bag which had the colours of the Nigerian flag.\nThe 47 year old was one of the guests invited to an event by media mogul, Debola Lagos.\nSee more photos:",
+    "summary": "Kate Henshaw Veteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.\nShe flaunted her hand bag which had the colours of the Nigerian flag.\nThe 47 year old was one of the guests invited to an event by media mogul, Debola Lagos.",
     "meta_description": "Online portal that provide breaking news and updates in Nigeria",
     "meta_lang": "",
     "meta_favicon": "https://www.linknaija.com/favicon.ico",
     "meta_site_name": "LinkNaija | Nigeria's popular news platform",
     "canonical_link": "https://www.linknaija.com/2018/10/nollywood-actress-kate-henshaw-is.html",
-    "text": "Photos which were released online and are making the rounds on social media, show Nigerian actress, Kate Henshaw, looking absolutely stunning in new make-up photos.\n\nKate Henshaw\n\nVeteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.\n\nShe flaunted her hand bag which had the colours of the Nigerian flag. The 47 year old was one of the guests invited to an event by media mogul, Debola Lagos.\n\nSee more photos:",
-    "text_cleaned": "Photos which were released online and are making the rounds on social media, show Nigerian actress, Kate Henshaw, looking absolutely stunning in new make-up photos.\n\nKate Henshaw\n\nVeteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.\n\nShe flaunted her hand bag which had the colours of the Nigerian flag. The 47 year old was one of the guests invited to an event by media mogul, Debola Lagos.\n\nSee more photos:"
+    "text": "Kate Henshaw\n\nVeteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.\n\nShe flaunted her hand bag which had the colours of the Nigerian flag. The 47 year old was one of the guests invited to an event by media mogul, Debola Lagos.",
+    "text_cleaned": "Kate Henshaw\n\nVeteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.\n\nShe flaunted her hand bag which had the colours of the Nigerian flag. The 47 year old was one of the guests invited to an event by media mogul, Debola Lagos."
 }
