+++ tests/data/metadata/cnn_article.json
@@ -1,5 +1,5 @@
 {
-    "url": "https://edition.cnn.com/2013/11/27/travel/weather-thanksgiving/index.html",
+    "url": "http://www.cnn.com/2013/11/27/travel/weather-thanksgiving/index.html",
     "read_more_link": "",
     "language": "en",
     "title": "After storm, forecasters see smooth sailing for Thanksgiving",
