+++ tests/data/txt/article_with_br.txt
@@ -1,5 +1,3 @@
-Puppet Masters
-
 An exclusive investigation by The Grayzone reveals new details on the critical role Sheldon Adelson's Las Vegas Sands played in an apparent CIA spying operation targeting Julian Assange, and exposes the Sands security staff who helped coordinate the malicious campaign.
 
 "I was the CIA director. We lied, we cheated, we stole."
