+++ tests/data/txt/article_with_divs.txt
@@ -1,9 +1,5 @@
-Photos which were released online and are making the rounds on social media, show Nigerian actress, Kate Henshaw, looking absolutely stunning in new make-up photos.
-
 Kate Henshaw
 
 Veteran Nollywood actress, Kate Henshaw, who in 2008, won the Africa Movie Academy Award for Best Actress in a Leading Role, has taken to her Instagram page to share some stunning photos of herself.
 
 She flaunted her hand bag which had the colours of the Nigerian flag. The 47 year old was one of the guests invited to an event by media mogul, Debola Lagos.
-
-See more photos:
