+++ tests/data/txt/cnn_001.txt
@@ -1,5 +1,3 @@
-CNN —
-
 Justice Samuel Alito is the tip of the spear for conservatives challenging the Biden administration during oral arguments at the Supreme Court.
 
 He’s a fierce questioner, ready to trap advocates in their arguments. And he becomes demonstrably riled when he fails to get the answer he wants. He shakes his head and rolls his eyes.
