+++ tests/data/txt/cnn_002.txt
@@ -1,5 +1,3 @@
-New York CNN —
-
 Donald Trump brought bombastic rhetoric to the witness stand Monday in the civil fraud case against him and his business, as he spent his time on the stand attacking the New York attorney general who brought the case and the judge overseeing the trial itself.
 
 Trump’s testimony at times mimicked his appearances on the campaign trail, where the former president has made the four criminal cases against him – along with the New York attorney general’s civil fraud case – a central part of his argument to be elected president again in 2024.
