+++ tests/data/txt/fox13now_001.txt
@@ -1,5 +1,3 @@
-By Leigh Ann Caldwell
-
 WASHINGTON (CNN) — Not everyone subscribes to a New Year’s resolution, but Americans will be required to follow new laws in 2014.
 
 Some 40,000 measures taking effect range from sweeping, national mandates under Obamacare to marijuana legalization in Colorado, drone prohibition in Illinois and transgender protections in California.
