+++ tests/data/txt/japanese_article.txt
