+++ tests/data/txt/wired_001.txt
@@ -10,8 +10,6 @@ The flagship component of the West Bank’s surveillance infrastructure is known
 
 Another strand, known as Blue Wolf, has been described as “Facebook for Palestinians.” It requires Israeli soldiers to photograph Palestinians individually via a smartphone app in order to record them in the database. According to Breaking the Silence, an NGO made up of former Israeli soldiers that opposes Israeli military occupation of Palestinian territories, prizes were offered to different units based on how many Palestinians they could photograph within a week.
 
-A person crossing the street in Huwara, West Bank on November 04, 2023.Photograph: Dan Kitwood/Getty Images
-
 “There are indications that the data gathered by this software impacts whether a person can get a work permit, whether a person can move from place to place within the West Bank, whether a person can get into Israel and work there or leave the country via Ben Gurion Airport, and a whole range of other things,” says Antony Loewenstein, author of The Palestine Laboratory, a book investigating the links between Israel’s military and technology sector.
 
 According to Azza, as he approaches a checkpoint and is picked up by a facial recognition camera, Israeli soldiers can see his profile from up to 100 meters away. They know who he is, who his family members are, where he lives, and his entire history of interactions with them. When he was 16, Azza was arrested for allegedly having a knife, but he denied the allegations and was later found to be innocent. He says this data is displayed every time he passes a checkpoint, meaning he is unfairly targeted and sometimes detained for up to three hours at a time. In recent years, he’s begun pre-emptively stripping to his underwear each time he enters a checkpoint, to avoid the lengthy process of being searched multiple times. “It’s a violation of our privacy,” he says.
