+++ tests/data/txt/yna_co_kr.txt
